Reorganize core utility headers
