+++ Src/Core/App/Application.h
@@ -18,7 +18,7 @@
 #include <string>
 #include <cstdint>
 
-#include "Core/Base.h"
+#include "Core/Util/Base.h"
 #include "Core/Event/EventBus.h"
 #include "Core/Event/ScopedConnection.h"
 #include "Core/App/Window.h"
