+++ Src/Core/Input/Action/InputAction.h
@@ -13,7 +13,7 @@
 #include <string>
 #include <unordered_map>
 #include <vector>
-#include "Core/Base.h"
+#include "Core/Util/Base.h"
 #include "Core/Input/Action/InputModifier.h"
 #include "Core/Input/Action/InputPatterns.h"
 #include "Core/Input/Action/InputSource.h"
