+++ Src/Demos/DemoRegistry.h
@@ -15,7 +15,7 @@
 #include <string>
 #include <vector>
 
-#include "Core/Base.h"
+#include "Core/Util/Base.h"
 
 class DemoBase;
