+++ Tests/Support/DiagnosticsTestSupport.h
@@ -6,7 +6,7 @@
 #include <thread>
 #include <vector>
 
-#include "Core/Base.h"
+#include "Core/Util/Base.h"
 #include "Core/Diagnostics/Logging/ImGuiConsoleSink.h"
 #include "Core/Diagnostics/Logging/Logger.h"
 #include "TestPaths.h"
