+++ Tests/Unit/TestLayerStack.cpp
@@ -3,7 +3,7 @@
 #include <string>
 
 #include "Core/App/LayerStack.h"
-#include "Core/Base.h"
+#include "Core/Util/Base.h"
 #include "TestLayer.h"
 
 TEST(LayerStackTests, PushLayerKeepsNormalLayerOrder)
