+++ Tests/Unit/TestLayerStackLifecycle.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include "Core/Base.h"
+#include "Core/Util/Base.h"
 #include "Core/App/LayerStack.h"
 #include "LifecycleTrackingLayer.h"
